+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Validation\Constraint::try()`
+
 ## 3.0.0 - 2026-01-18
 
 ### Changed